+++ push.sh
@@ -219,8 +219,12 @@ if [ $1 = "room/intro" -o \
         path="$assets/$1/content"
         mkdir -p "$path"
 
-        cp "$source"/*.png "$path"
         cp "$source"/*.mp4 "$path"
+
+        if exists "$source"/*.png; then
+
+            cp "$source"/*.png "$path"
+        fi
 
         if exists "$source"/*.jpg; then
